type-racer/requirements.txt
fastapi
uvicorn[standard]
aiofiles
firebase-admin
python-dotenv
